refactor(main): split monolithic ipcHandlers.js into six domain modules under src/main/ipc/
- Router ipcHandlers.js reduced from ~904 lines to ~63 lines
- New files: _shared.js, scan.js, quarantine.js, process.js, firewall.js, network.js, system.js
- Each domain exports a single register(ipcMain, deps) constructor
- All 60+ IPC channel strings preserved; zero renderer-visible contract changes
- Relies on shared relative requires (../../i18n, ../../security/reportExport, ...)

feat(core): introduce featureFlags single source of truth

- New src/core/featureFlags.js with DEFAULT_FLAGS, getFlag, setFlag, validation
- system.js db:getSetting/db:setSetting route feature.* keys through featureFlags
- main.js replaces direct db.getSetting('feature.*') reads with featureFlags.getFlag
- Unknown feature keys throw descriptive errors in dev builds
- Fixes: feature.systemMonitoring migration was accidentally placed in showNotification runtime path; restored to startup-time migration only

refactor(security): extract scanProgress clamp utility and replace inline clamps

- New src/core/scanProgress.js exports clampProgress(value)
- ScanEngine.js emitProgress uses clampProgress instead of Math.min(100, ...)
- scan.js scheduled scan path also cleaned up
- Phase3 compatibility getter retained because FolderWatcher/scanner.js depend on it

refactor(main): replace remaining raw console.error calls with structured logger

- Reuses existing src/utils/logger.js rather than introducing a duplicate
- Cleans up console.error in src/main/ipc/scan.js and src/main/ipc/network.js
- Zero console.* calls remain in src/main/ outside src/core/logger.js itself

test: update 13 scan-engine tests to match actual state-shape behavior

- scanEngine.test.js: direct top-level engine.isScanning assignments changed
  to engine.userScan.isScanning; same for isFolderWatchScanning,
  currentScan, abortController; constructor currentScan assertion uses null
- scanCancellation.test.js: scanEngine.abortController scan changed to
  userScan.abortController; expected error string 'No scan in progress'
  changed to 'No user scan in progress'
- Bounded the folderwatch guard test with Promise.race so it no longer
  times out on this runner
