+++ src/main/ipc/_shared.js
@@ -1,0 +1,23 @@
+const https = require('https');
+
+function requestText(url, options = {}) {
+  return new Promise((resolve, reject) => {
+    const req = https.request(url, {
+      method: 'GET',
+      headers: {
+        'User-Agent': 'Soterios',
+        ...options.headers,
+      },
+    }, (res) => {
+      let body = '';
+      res.setEncoding('utf8');
+      res.on('data', chunk => { body += chunk; });
+      res.on('end', () => resolve({ statusCode: res.statusCode, body }));
+    });
+    req.on('error', reject);
+    req.setTimeout(15000, () => req.destroy(new Error('Request timed out')));
+    req.end();
+  });
+}
+
+module.exports = { requestText };
