+++ src/main/ipc/quarantine.js
@@ -1,0 +1,13 @@
+const { ipcMain } = require('electron');
+
+function register(mainWindow, { quarantineManager }) {
+  ipcMain.handle('quarantine:restore', async (_event, id) => {
+    return quarantineManager.restore(id);
+  });
+
+  ipcMain.handle('quarantine:delete', async (_event, id) => {
+    return quarantineManager.delete(id);
+  });
+}
+
+module.exports = { register };
